requirements.txt
appdirs==1.4.3
asn1crypto==0.22.0
cffi==1.10.0
coreapi==2.3.0
coreschema==0.0.4
cryptography==1.9
Django>1.11
django-appconf==1.0.2
django-crispy-forms==1.6.1
django-cruds-adminlte
django-filter==1.0.2
django-image-cropping==1.0.4
Django-Select2==5.9.0
djangoajax==2.3.7
djangorestframework==3.6.2
easy-thumbnails==2.4.1
idna==2.5
itypes==1.1.0
Jinja2==2.9.6
Markdown==2.6.8
MarkupSafe==1.0
olefile==0.44
packaging==16.8
Pillow==4.1.0
pycparser==2.17
pycrypto==2.6.1
Pygments==2.2.0
pyOpenSSL==16.2.0
pyparsing==2.2.0
pytz==2017.2
requests==2.13.0
six==1.10.0
uritemplate==3.0.0
dateutils
pyfva
pycryptodome
